Add basic support for pictures
Add basic support for pictures posted to Telegram. Now it just
only pass the telegram url to irc. That url is valid only for 1h.
Next step is to make possible to save it to local http server
and send link to there. That should be optional as not everybody
have own server to run this.

Now the program crash if photo is send to telegram before
irc part is fully connected. We should consider to wait that before
accepting any messages from telegram.