Add basic support for multiserver (fixes: https://github.com/FruitieX/teleirc/issues/190)
This adds support for basic muliserver setup.

This does not have support for different ports, nick or other server
configurations just yet. This only sets up basic muliserver setup using
one common ircOptions.

This does not modify the config structure, it only adds the option to
set server in config.channels, if ircServer is not set in
config.channels entrys it will default to config.ircServer.